Rename HarmonizationDataContainer -> CatalogueHarmonizationPanel
