+++ src/components/CatalogueHarmonizationPanel.vue
